+++ interfaces/src/main/java/ar/edu/itba/paw/persistence/UserDao.java
@@ -10,6 +10,6 @@ public interface UserDao {
 
     Optional<User> getUserByEmail(String email);
 
-    Optional<User> getUserById(long id);
+    Optional<User> getUserById(int id);
     Optional<User> getUserById(String id);
 }
