+++ webapp/src/main/java/ar/edu/itba/paw/webapp/validators/interfaces/ImageConstraint.java
@@ -19,5 +19,5 @@ public @interface ImageConstraint {
     Class<?>[] groups() default {};
     Class<? extends Payload>[] payload() default {};
 
-    long maxSize();
+    int maxSize();
 }
